dtn-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dtn-mcp"
version = "0.1.0"
description = "MCP server for the DTN Test Bed (Mac API proxy + CSV export)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.2.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
dtn-mcp = "dtn_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["dtn_mcp"]

[tool.hatch.build.targets.sdist]
include = ["dtn_mcp", "README.md", "pyproject.toml"]
